EASY-1731 easy-validate-dans-bag fails prematurely on malformed bag-info.txt
